requirements.txt
BM25
faiss-cpu
google-genai
groq
huggingface_hub
ipywidgets
langchain
langchain-google-genai
langchain-groq
langchain-huggingface
langchain-openai
numpy
openai
pydantic
pypdf
pytest
python-dotenv
rich
sentence-transformers

